+++ src/main/java/com/pulse/content/ContentApplication.java
@@ -3,9 +3,11 @@ package com.pulse.content;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+import org.springframework.scheduling.annotation.EnableAsync;
 
 @SpringBootApplication
-@EnableDiscoveryClient
+@EnableDiscoveryClient // Eureka 서버에 서비스를 등록하기 위한 어노테이션
+@EnableAsync           // 비동기 처리를 위한 어노테이션
 public class ContentApplication {
 
     public static void main(String[] args) {
